ExpenseTracker/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
python-dateutil==2.8.2
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
